+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/AlternativeHeadquarterRelationUpsertService.kt
@@ -21,9 +21,9 @@ package org.eclipse.tractusx.bpdm.pool.service
 
 import org.eclipse.tractusx.bpdm.pool.api.model.RelationType
 import org.eclipse.tractusx.bpdm.pool.dto.UpsertResult
-import org.eclipse.tractusx.bpdm.pool.dto.UpsertType
 import org.eclipse.tractusx.bpdm.pool.entity.LegalEntityDb
 import org.eclipse.tractusx.bpdm.pool.entity.RelationDb
+import org.eclipse.tractusx.bpdm.pool.exception.BpdmValidationException
 import org.eclipse.tractusx.bpdm.pool.repository.RelationRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -38,13 +38,11 @@ class AlternativeHeadquarterRelationUpsertService(
     override fun upsertRelation(upsertRequest: IRelationUpsertStrategyService.UpsertRequest): UpsertResult<RelationDb>{
         val standardisedRequest = standardise(upsertRequest)
 
+        validateOnlyOneAlternative(standardisedRequest)
+
         val result = relationUpsertService.upsertRelation(
             RelationUpsertService.UpsertRequest(standardisedRequest.source, standardisedRequest.target, RelationType.IsAlternativeHeadquarterFor)
         )
-
-        if(result.upsertType == UpsertType.Created){
-            createTransitiveRelations(standardisedRequest)
-        }
 
         return result
     }
@@ -63,29 +61,22 @@ class AlternativeHeadquarterRelationUpsertService(
         return upsertRequest
     }
 
-    private fun createTransitiveRelations(upsertRequest: IRelationUpsertStrategyService.UpsertRequest){
-        val transitiveSourceRelations = relationRepository.findInSourceOrTarget(RelationType.IsAlternativeHeadquarterFor, upsertRequest.source)
-        val transitiveTargetRelations = relationRepository.findInSourceOrTarget(RelationType.IsAlternativeHeadquarterFor, upsertRequest.target)
-
-        val transitiveSourceRequests = createTransitiveUpsertRequests(upsertRequest.target, transitiveSourceRelations)
-        val transitiveTargetRequests = createTransitiveUpsertRequests(upsertRequest.source, transitiveTargetRelations)
-
-        transitiveSourceRequests
-            .plus(transitiveTargetRequests)
-            .filter { it.source.id != upsertRequest.source.id || it.target.id != upsertRequest.target.id }
-            .distinctBy { Pair(it.source.id, it.target.id) }
-            .map { RelationUpsertService.UpsertRequest(it.source, it.target, RelationType.IsAlternativeHeadquarterFor) }
-            .forEach { relationUpsertService.upsertRelation(it) }
+    private fun validateOnlyOneAlternative(upsertRequest: IRelationUpsertStrategyService.UpsertRequest){
+        validateOnlyOneAlternative(upsertRequest.source, upsertRequest)
+        validateOnlyOneAlternative(upsertRequest.target, upsertRequest)
     }
 
-    private fun createTransitiveUpsertRequests(legalEntity: LegalEntityDb, relations: Set<RelationDb>): List<IRelationUpsertStrategyService.UpsertRequest>{
-        val allLegalEntities = relations.flatMap { listOf(it.startNode, it.endNode) }.toSet()
+    private fun validateOnlyOneAlternative(legalEntity: LegalEntityDb, upsertRequest: IRelationUpsertStrategyService.UpsertRequest){
+        val relations =  relationRepository.findInSourceOrTarget(RelationType.IsAlternativeHeadquarterFor, legalEntity)
+        val otherRelations =  relations.filterNot { it.startNode.bpn == upsertRequest.source.bpn && it.endNode.bpn == upsertRequest.target.bpn }
 
-        return  allLegalEntities
-            .filter { legalEntity.id != it.id }
-            .map { IRelationUpsertStrategyService.UpsertRequest(legalEntity, it) }
-            .map { standardise(it) }
-            .distinctBy { Pair(it.source.id, it.target.id) }
+        if(otherRelations.isNotEmpty()){
+            val otherRelation = otherRelations.first()
+            val otherLegalEntity = if(otherRelation.startNode.bpn == legalEntity.bpn) otherRelation.endNode else otherRelation.startNode
+            throw BpdmValidationException(
+                "Invalid 'IsAlternativeHeadquarter' relation: Legal Entity ${legalEntity.bpn} is already alternative headquarter to ${otherLegalEntity.bpn}"
+            )
+        }
     }
 
 }