+++ CHANGELOG.md
@@ -8,6 +8,10 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 
 ## [7.2.0] - tbd
 
+### Changed
+
+- BPDM Pool: each legal entity can now have only one alternative headquarter
+
 ## [7.1.0] - 2025-09-30
 
 ### Added